requirements.txt
chromadb
openai
python-dotenv
streamlit
